+++ src/main/ts/ast/Exports.ts
@@ -1,18 +1,35 @@
 import * as estree from 'estree';
 
+export enum ExportInfoKind {
+  Specified,
+  Variable
+}
+
 interface ExportInfo {
   name: string;
   fromName: string;
+  kind: ExportInfoKind;
 }
 
-const create = (name: string, fromName: string): ExportInfo => {
-  return { name, fromName };
+const isVariableDeclaration = (declaration: estree.Declaration): declaration is estree.VariableDeclaration => {
+  return declaration && declaration.type === 'VariableDeclaration';
+};
+
+const create = (name: string, fromName: string, kind: ExportInfoKind): ExportInfo => {
+  return { name, fromName, kind };
 };
 
 const readExportDeclaration = (node: estree.ExportNamedDeclaration): ExportInfo[] => {
-  return node.specifiers.map((specifier) => {
-    return create(specifier.exported.name, specifier.local.name);
-  });
+  if (isVariableDeclaration(node.declaration)) {
+    return node.declaration.declarations.map((declaration) => {
+      const id = declaration.id as estree.Identifier;
+      return create(id.name, id.name, ExportInfoKind.Variable);
+    });
+  } else {
+    return node.specifiers.map((specifier) => {
+      return create(specifier.exported.name, specifier.local.name, ExportInfoKind.Specified);
+    });
+  }
 };
 
 const readExports = (program: estree.Program): ExportInfo[] => {